assignment_5/requirements.txt
more_itertools
elasticsearch
psycopg2-binary